pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pptxtpl"
version = "0.2.0"
description = "Jinja2 templating for PowerPoint .pptx files"
readme = "README.md"
license = "MIT"
requires-python = ">=3.9"
dependencies = [
    "python-pptx",
    "Jinja2",
    "lxml",
]
